Move thread_pool_impl to a separate library
