+++ base/BUILD
@@ -23,11 +23,21 @@ cc_library(
 
 cc_library(
     name = "thread_pool",
-    srcs = ["thread_pool.cc"],
     hdrs = [
         "thread_pool.h",
         "function_ref.h",
     ],
+    visibility = ["//visibility:public"],
+)
+
+cc_library(
+    name = "thread_pool_impl",
+    srcs = ["thread_pool_impl.cc"],
+    hdrs = [
+        "thread_pool_impl.h",
+        "function_ref.h",
+    ],
+    deps = [":thread_pool"],
     visibility = ["//visibility:public"],
 )
