+++ runtime/test/evaluate_benchmark.cc
@@ -5,7 +5,7 @@
 #include <cstdint>
 #include <chrono>
 
-#include "base/thread_pool.h"
+#include "base/thread_pool_impl.h"
 #include "runtime/evaluate.h"
 #include "runtime/expr.h"
